+++ src/AxisSweep3.h
@@ -12,9 +12,9 @@ namespace BulletSharp
 		AxisSweep3(btAxisSweep3* axisSweep);
 
 	public:
-		AxisSweep3(Vector3 worldAabbMin, Vector3 worldAabbMax, int maxHandles, BulletSharp::OverlappingPairCache^ pairCache, bool disableRaycastAccelerator);
-		AxisSweep3(Vector3 worldAabbMin, Vector3 worldAabbMax, int maxHandles, BulletSharp::OverlappingPairCache^ pairCache);
-		AxisSweep3(Vector3 worldAabbMin, Vector3 worldAabbMax, int maxHandles);
+		AxisSweep3(Vector3 worldAabbMin, Vector3 worldAabbMax, unsigned short maxHandles, BulletSharp::OverlappingPairCache^ pairCache, bool disableRaycastAccelerator);
+		AxisSweep3(Vector3 worldAabbMin, Vector3 worldAabbMax, unsigned short maxHandles, BulletSharp::OverlappingPairCache^ pairCache);
+		AxisSweep3(Vector3 worldAabbMin, Vector3 worldAabbMax, unsigned short maxHandles);
 		AxisSweep3(Vector3 worldAabbMin, Vector3 worldAabbMax);
 	};
 
@@ -24,9 +24,9 @@ namespace BulletSharp
 		AxisSweep3_32Bit(bt32BitAxisSweep3* axisSweep);
 
 	public:
-		AxisSweep3_32Bit(Vector3 worldAabbMin, Vector3 worldAabbMax, int maxHandles, BulletSharp::OverlappingPairCache^ pairCache, bool disableRaycastAccelerator);
-		AxisSweep3_32Bit(Vector3 worldAabbMin, Vector3 worldAabbMax, int maxHandles, BulletSharp::OverlappingPairCache^ pairCache);
-		AxisSweep3_32Bit(Vector3 worldAabbMin, Vector3 worldAabbMax, int maxHandles);
+		AxisSweep3_32Bit(Vector3 worldAabbMin, Vector3 worldAabbMax, unsigned int maxHandles, BulletSharp::OverlappingPairCache^ pairCache, bool disableRaycastAccelerator);
+		AxisSweep3_32Bit(Vector3 worldAabbMin, Vector3 worldAabbMax, unsigned int maxHandles, BulletSharp::OverlappingPairCache^ pairCache);
+		AxisSweep3_32Bit(Vector3 worldAabbMin, Vector3 worldAabbMax, unsigned int maxHandles);
 		AxisSweep3_32Bit(Vector3 worldAabbMin, Vector3 worldAabbMax);
 	};
 };
